Bump karaf to 4.3 and jackson to 2.11.3

+++ core/src/test/java/org/apache/brooklyn/util/core/internal/TypeCoercionsTest.java
@@ -66,64 +66,64 @@ public class TypeCoercionsTest {
     
     @Test
     public void testCoerceStringToPrimitive() {
-        assertEquals(TypeCoercions.coerce("1", Character.class), (Character)'1');
-        assertEquals(TypeCoercions.coerce(" ", Character.class), (Character)' ');
-        assertEquals(TypeCoercions.coerce("1", Short.class), (Short)((short)1));
-        assertEquals(TypeCoercions.coerce("1", Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce("1", Long.class), (Long)1l);
-        assertEquals(TypeCoercions.coerce("1", Float.class), 1f);
-        assertEquals(TypeCoercions.coerce("1", Double.class), 1d);
-        assertEquals(TypeCoercions.coerce("true", Boolean.class), (Boolean)true);
-        assertEquals(TypeCoercions.coerce("False", Boolean.class), (Boolean)false);
-        assertEquals(TypeCoercions.coerce("true ", Boolean.class), (Boolean)true);
+        assertEquals(TypeCoercions.coerce("1", Character.class), (Character) '1');
+        assertEquals(TypeCoercions.coerce(" ", Character.class), (Character) ' ');
+        assertEquals(TypeCoercions.coerce("1", Short.class), (Short) ((short) 1));
+        assertEquals(TypeCoercions.coerce("1", Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce("1", Long.class), (Long) 1l);
+        assertEquals(TypeCoercions.coerce("1", Float.class), (Float) 1f);
+        assertEquals(TypeCoercions.coerce("1", Double.class), (Double) 1d);
+        assertEquals(TypeCoercions.coerce("true", Boolean.class), (Boolean) true);
+        assertEquals(TypeCoercions.coerce("False", Boolean.class), (Boolean) false);
+        assertEquals(TypeCoercions.coerce("true ", Boolean.class), (Boolean) true);
         assertNull(TypeCoercions.coerce(null, Boolean.class), null);
 
-        assertEquals(TypeCoercions.coerce("1", char.class), (Character)'1');
-        assertEquals(TypeCoercions.coerce("1", short.class), (Short)((short)1));
-        assertEquals(TypeCoercions.coerce("1", int.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce("1", long.class), (Long)1l);
-        assertEquals(TypeCoercions.coerce("1", float.class), 1f);
-        assertEquals(TypeCoercions.coerce("1", double.class), 1d);
-        assertEquals(TypeCoercions.coerce("TRUE", boolean.class), (Boolean)true);
-        assertEquals(TypeCoercions.coerce("false", boolean.class), (Boolean)false);
+        assertEquals(TypeCoercions.coerce("1", char.class), (Character) '1');
+        assertEquals(TypeCoercions.coerce("1", short.class), (Short) ((short) 1));
+        assertEquals(TypeCoercions.coerce("1", int.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce("1", long.class), (Long) 1l);
+        assertEquals(TypeCoercions.coerce("1", float.class), (Float) 1f);
+        assertEquals(TypeCoercions.coerce("1", double.class), (Double) 1d);
+        assertEquals(TypeCoercions.coerce("TRUE", boolean.class), (Boolean) true);
+        assertEquals(TypeCoercions.coerce("false", boolean.class), (Boolean) false);
     }
 
     @Test
     public void testCoercePrimitivesToSameType() {
-        assertEquals(TypeCoercions.coerce('1', Character.class), (Character)'1');
-        assertEquals(TypeCoercions.coerce((short)1, Short.class), (Short)((short)1));
-        assertEquals(TypeCoercions.coerce(1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce(1l, Long.class), (Long)1l);
-        assertEquals(TypeCoercions.coerce(1f, Float.class), 1f);
-        assertEquals(TypeCoercions.coerce(1d, Double.class), 1d);
-        assertEquals(TypeCoercions.coerce(true, Boolean.class), (Boolean)true);
+        assertEquals(TypeCoercions.coerce('1', Character.class), (Character) '1');
+        assertEquals(TypeCoercions.coerce((short) 1, Short.class), (Short) ((short) 1));
+        assertEquals(TypeCoercions.coerce(1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce(1l, Long.class), (Long) 1l);
+        assertEquals(TypeCoercions.coerce(1f, Float.class), (Float) 1f);
+        assertEquals(TypeCoercions.coerce(1d, Double.class), (Double) 1d);
+        assertEquals(TypeCoercions.coerce(true, Boolean.class), (Boolean) true);
     }
     
     @Test
     public void testCastPrimitives() {
-        assertEquals(TypeCoercions.coerce(1L, Character.class), (Character)(char)1);
-        assertEquals(TypeCoercions.coerce(1L, Byte.class), (Byte)(byte)1);
-        assertEquals(TypeCoercions.coerce(1L, Short.class), (Short)(short)1);
-        assertEquals(TypeCoercions.coerce(1L, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce(1L, Long.class), (Long)(long)1);
-        assertEquals(TypeCoercions.coerce(1L, Float.class), (float)1);
-        assertEquals(TypeCoercions.coerce(1L, Double.class), (double)1);
-        
-        assertEquals(TypeCoercions.coerce(1L, char.class), (Character)(char)1);
-        assertEquals(TypeCoercions.coerce(1L, byte.class), (Byte)(byte)1);
-        assertEquals(TypeCoercions.coerce(1L, short.class), (Short)(short)1);
-        assertEquals(TypeCoercions.coerce(1L, int.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce(1L, long.class), (Long)(long)1);
-        assertEquals(TypeCoercions.coerce(1L, float.class), (float)1);
-        assertEquals(TypeCoercions.coerce(1L, double.class), (double)1);
-        
-        assertEquals(TypeCoercions.coerce((char)1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce((byte)1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce((short)1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce(1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce((long)1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce((float)1, Integer.class), (Integer)1);
-        assertEquals(TypeCoercions.coerce((double)1, Integer.class), (Integer)1);
+        assertEquals(TypeCoercions.coerce(1L, Character.class), (Character) (char) 1);
+        assertEquals(TypeCoercions.coerce(1L, Byte.class), (Byte) (byte) 1);
+        assertEquals(TypeCoercions.coerce(1L, Short.class), (Short) (short) 1);
+        assertEquals(TypeCoercions.coerce(1L, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce(1L, Long.class), (Long) (long) 1);
+        assertEquals(TypeCoercions.coerce(1L, Float.class), (Float) (float) 1);
+        assertEquals(TypeCoercions.coerce(1L, Double.class), (Double) (double) 1);
+
+        assertEquals(TypeCoercions.coerce(1L, char.class), (Character) (char) 1);
+        assertEquals(TypeCoercions.coerce(1L, byte.class), (Byte) (byte) 1);
+        assertEquals(TypeCoercions.coerce(1L, short.class), (Short) (short) 1);
+        assertEquals(TypeCoercions.coerce(1L, int.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce(1L, long.class), (Long) (long) 1);
+        assertEquals(TypeCoercions.coerce(1L, float.class), (Float) (float) 1);
+        assertEquals(TypeCoercions.coerce(1L, double.class), (Double) (double) 1);
+
+        assertEquals(TypeCoercions.coerce((char) 1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce((byte) 1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce((short) 1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce(1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce((long) 1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce((float) 1, Integer.class), (Integer) 1);
+        assertEquals(TypeCoercions.coerce((double) 1, Integer.class), (Integer) 1);
     }
     
     @Test
